Move optimizer known btypes to their own file and rename
